+++ blackbox.php
@@ -8,28 +8,25 @@ use Innmind\BlackBox\{
     Runner\Load,
     Runner\CodeCoverage,
     PHPUnit,
+    Prove,
 };
 
 Application::new($argv)
-    ->when(
-        \getenv('BLACKBOX_SET_SIZE') !== false,
-        static fn(Application $app) => $app->scenariiPerProof((int) \getenv('BLACKBOX_SET_SIZE')),
-    )
-    ->when(
-        \getenv('ENABLE_COVERAGE') !== false,
-        static fn(Application $app) => $app
+    ->map(static fn($app) => match (\getenv('BLACKBOX_ENV')) {
+        'extensive' => $app->scenariiPerProof(1_000),
+        'coverage' => $app
             ->codeCoverage(
                 CodeCoverage::of(
                     __DIR__.'/src/',
                     __DIR__.'/tests/',
                     __DIR__.'/proofs/',
                 )
-                    ->dumpTo('coverage.clover')
-                    ->enableWhen(true),
+                    ->dumpTo('coverage.clover'),
             ),
-    )
-    ->tryToProve(static function() {
+        default => $app,
+    })
+    ->tryToProve(static function(Prove $prove) {
         yield from PHPUnit\Load::testsAt(__DIR__.'/tests/');
-        yield from Load::everythingIn(__DIR__.'/proofs/')();
+        yield from Load::everythingIn(__DIR__.'/proofs/')($prove);
     })
     ->exit();